Added WL NB for session 1
Wolfram Language APIs (APIFunction[ ]) used with JavaScript ReactJS libraries (create-react-app).  This first session introduces the motivations, requirements, and starts with very simple examples.  Each example is subsequently layered on with more considerations and requirements.  Session one concludes with calling for and returning a JSON object processed by JS and displayed in a browser.